Add comentary to all hpp
